% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_aa.R
\name{aa}
\alias{aa}
\title{Build an amino acid (AA) vector}
\usage{
aa(...)
}
\arguments{
\item{...}{character to turn into AA. Can be a set of name-value pairs.}
}
\value{
vector of class bioseq_aa
}
\description{
\code{aa()} build a AA vector from a character vector.
}
\examples{

aa("AGGTGC", "TTCGA")

aa(Seq_1 = "AGGTGC", Seq_2 = "TTCGA")

x <- c("AGGTGC", "TTCGA")
aa(x)

}
\seealso{
Other classes: 
\code{\link{dna}()},
\code{\link{rna}()}
}
\concept{classes}
