% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_aa.R
\name{as_aa}
\alias{as_aa}
\title{Coercion to an amino acid (AA) vector}
\usage{
as_aa(x)
}
\arguments{
\item{x}{An object to coerce.}
}
\value{
An amino acid vector of class bioseq_aa
}
\description{
Coercion to an amino acid (AA) vector
}
