% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_DNAbin.R
\name{as_AAbin}
\alias{as_AAbin}
\title{Coerce to AAbin}
\usage{
as_AAbin(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other parameters.}
}
\value{
An AAbin object.
}
\description{
Coerce to AAbin
}
\seealso{
Other conversions: \code{\link{as-tibble-ape}},
  \code{\link{as-tibble-bioseq}}, \code{\link{as_DNAbin}},
  \code{\link{as_aa}}, \code{\link{as_dna}},
  \code{\link{as_rna}}, \code{\link{as_seqinr_alignment}}
}
\concept{conversions}
