% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_dna.R
\name{as_dna}
\alias{as_dna}
\title{Coercion to DNA vector}
\usage{
as_dna(x)
}
\arguments{
\item{x}{An object to coerce.}
}
\value{
A DNA vector of class bioseq_dna
}
\description{
Coercion to DNA vector
}
\seealso{
Other conversions: \code{\link{as-tibble-ape}},
  \code{\link{as-tibble-bioseq}}, \code{\link{as_AAbin}},
  \code{\link{as_DNAbin}}, \code{\link{as_aa}},
  \code{\link{as_rna}}, \code{\link{as_seqinr_alignment}}
}
\concept{conversions}
