% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_DNAbin.R
\name{as_DNAbin}
\alias{as_DNAbin}
\title{Coerce to DNAbin}
\usage{
as_DNAbin(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other parameters.}
}
\value{
A DNAbin object.
}
\description{
Coerce to DNAbin
}
