% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_DNAbin.R
\name{as_DNAbin.tbl_df}
\alias{as_DNAbin.tbl_df}
\title{Coerce tibble to DNAbin}
\usage{
\method{as_DNAbin}{tbl_df}(x, sequences, labels = NULL, ...)
}
\arguments{
\item{x}{a tibble.}

\item{sequences}{Name of the tibble column that stores the sequences.}

\item{labels}{Name of the tibble column that stores the sequence labels.}

\item{...}{Other params.}
}
\value{
A DNAbin object.
}
\description{
Coerce tibble to DNAbin
}
