% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_rna.R
\name{as_rna}
\alias{as_rna}
\title{Coercion to RNA vector}
\usage{
as_rna(x)
}
\arguments{
\item{x}{An object to coerce.}
}
\value{
A RNA vector of class bioseq_rna
}
\description{
Coercion to RNA vector
}
\seealso{
Other conversions: \code{\link{as-tibble-ape}},
  \code{\link{as-tibble-bioseq}}, \code{\link{as_AAbin}},
  \code{\link{as_DNAbin}}, \code{\link{as_aa}},
  \code{\link{as_dna}}, \code{\link{as_seqinr_alignment}}
}
\concept{conversions}
