% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_seqinr.R
\name{as_seqinr_alignment}
\alias{as_seqinr_alignment}
\title{Coerce to seqinr alignment}
\usage{
as_seqinr_alignment(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other parameters.}
}
\value{
An alignment object.
}
\description{
Coerce to seqinr alignment
}
