% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioseq-package.R
\docType{package}
\name{bioseq-package}
\alias{bioseq}
\alias{bioseq-package}
\title{bioseq: A Toolbox for Manipulating Biological Sequences}
\description{
Classes and functions to work with biological sequences (DNA, RNA and amino acid sequences).
    Implements S3 infrastructure to work with biological sequences.
    Provides a collection of functions to perform biological conversion among classes
    (transcription, translation) and basic operations on sequences
    (detection, selection and replacement based on positions or patterns).
    The package also provides functions to import and export sequences from and to other package formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fkeck.github.io/bioseq/}
  \item Report bugs at \url{https://github.com/fkeck/bioseq/issues}
}

}
\author{
\strong{Maintainer}: Francois Keck \email{francois.keck@gmail.com} (0000-0002-3323-4167) [copyright holder]

}
\keyword{internal}
