% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-arg.R
\name{bioseq}
\alias{bioseq}
\title{Interactive Bio Sequence}
\usage{
bioseq()
}
\value{
Interactive function gives DNA complement, Reverse complement or Transcribed Sequence or more detailed Information based on the option passed.
}
\description{
Basic Bioseq interactive function
}
\examples{
bioseq()
}
\author{
Dilip
}
