% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-arg.R
\name{bioseq_complement}
\alias{bioseq_complement}
\title{Complement of the Sequence}
\usage{
bioseq_complement(seq)
}
\arguments{
\item{seq}{The sequence in a character}
}
\value{
Character of the complement sequence
}
\description{
DNA sequence complement
}
\examples{
bioseq_complement("ATGCGATGTCGAGAGCAGCGATGC")
# Function will not be affected by case of the characters,
# return will be of all Upcased characters.
bioseq_complement("agtGTGacGCT")
}
\author{
Dilip
}
