% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-arg.R
\name{bioseq_rcomplement}
\alias{bioseq_rcomplement}
\title{Reverse Complement}
\usage{
bioseq_rcomplement(seq)
}
\arguments{
\item{seq}{The sequence in a character}
}
\value{
Reverse complement for input sequence
}
\description{
Reverse complement for a sequence
}
\examples{
bioseq_rcomplement("ATGAGTGCGTGCT")
bioseq_rcomplement("AGagaTgcAgATGCacgca")
}
\author{
Dilip
}
