% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-arg.R
\name{bioseq_transcribe}
\alias{bioseq_transcribe}
\title{Transcription of DNA}
\usage{
bioseq_transcribe(seq)
}
\arguments{
\item{seq}{The sequence in a character}
}
\value{
mRna sequence that is transcribed by input sequence.
}
\description{
Transcription of a DNA sequence
}
\examples{
bioseq_transcribe("AGTGACGAGCGAagcgaGACGGTGagagcg")
}
\author{
Dilip
}
