% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_codes.R
\name{dic_genetic_codes}
\alias{dic_genetic_codes}
\title{Genetic code tables}
\usage{
dic_genetic_codes()
}
\value{
A list of genetic code tables for DNA/RNA translation.
}
\description{
The function returns a list of named vectors with
Start, Stop and Full_name attributes.
}
