% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_dna.R
\name{dna}
\alias{dna}
\title{Build a DNA vector}
\usage{
dna(...)
}
\arguments{
\item{...}{characters to turn into DNA. Can be a set of name-value pairs.}
}
\value{
a vector of class \code{bioseq_dna}
}
\description{
\code{dna()} build a DNA vector from a character vector.
}
\examples{

dna("AGGTGC", "TTCGA")

dna(Seq_1 = "AGGTGC", Seq_2 = "TTCGA")

x <- c("AGGTGC", "TTCGA")
dna(x)

}
\seealso{
Other classes: 
\code{\link{aa}()},
\code{\link{rna}()}
}
\concept{classes}
