% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fragilaria}
\alias{fragilaria}
\title{DNA sequences (rbcL) for various Fragilaria}
\format{
A long character vector (unparsed FASTA).
}
\source{
GenBank \url{https://www.ncbi.nlm.nih.gov/genbank/} using
the following search term: "(rbcl) AND Fragilaria"
}
\usage{
fragilaria
}
\description{
An unparsed FASTA of DNA sequences (rbcL) for various strains of Fragilaria
retrieved from NCBI.
}
\seealso{
\code{\link{read_fasta}} to parse these data.
}
\keyword{datasets}
