% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_codes.R
\name{genetic-codes}
\alias{genetic-codes}
\title{Genetic code tables}
\description{
List of all genetic code tables available in \code{bioseq}.
The number in bold can be used to select a table in appropriate functions.
}
\section{Available genetic codes}{


\strong{1}\code{.    }Standard

\strong{2}\code{.    }Vertebrate Mitochondrial

\strong{3}\code{.    }Yeast Mitochondrial

\strong{4}\code{.    }Mold Mitochondrial; Protozoan Mitochondrial; Coelenterate Mitochondrial; Mycoplasma; Spiroplasma

\strong{5}\code{.    }Invertebrate Mitochondrial

\strong{6}\code{.    }Ciliate Nuclear; Dasycladacean Nuclear; Hexamita Nuclear

\strong{9}\code{.    }Echinoderm Mitochondrial; Flatworm Mitochondrial

\strong{10}\code{.    }Euplotid Nuclear

\strong{11}\code{.    }Bacterial, Archaeal and Plant Plastid

\strong{12}\code{.    }Alternative Yeast Nuclear

\strong{13}\code{.    }Ascidian Mitochondrial

\strong{14}\code{.    }Alternative Flatworm Mitochondrial

\strong{15}\code{.    }Blepharisma Macronuclear

\strong{16}\code{.    }Chlorophycean Mitochondrial

\strong{21}\code{.    }Trematode Mitochondrial

\strong{22}\code{.    }Scenedesmus obliquus Mitochondrial

\strong{23}\code{.    }Thraustochytrium Mitochondrial

\strong{24}\code{.    }Pterobranchia Mitochondrial

\strong{25}\code{.    }Candidate Division SR1 and Gracilibacteria

\strong{26}\code{.    }Pachysolen tannophilus Nuclear

\strong{27}\code{.    }Karyorelict Nuclear

\strong{28}\code{.    }Condylostoma Nuclear

\strong{29}\code{.    }Mesodinium Nuclear

\strong{30}\code{.    }Peritrich Nuclear

\strong{31}\code{.    }Blastocrithidia Nuclear

\strong{32}\code{.    }Balanophoraceae Plastid

\strong{33}\code{.    }Cephalodiscidae Mitochondrial
}

\references{
Andrzej (Anjay) Elzanowski and Jim Ostell at National Center for Biotechnology Information (NCBI), Bethesda, Maryland, U.S.A.
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=tgencodes}
}
