% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_aa.R, R/def_dna.R, R/def_rna.R
\name{vec_ptype2.bioseq_aa}
\alias{vec_ptype2.bioseq_aa}
\alias{vec_cast.bioseq_aa}
\alias{vec_ptype2.bioseq_dna}
\alias{vec_cast.bioseq_dna}
\alias{vec_ptype2.bioseq_rna}
\alias{vec_cast.bioseq_rna}
\title{Internal}
\usage{
\method{vec_ptype2}{bioseq_aa}(x, y, ...)

\method{vec_cast}{bioseq_aa}(x, to, ...)

\method{vec_ptype2}{bioseq_dna}(x, y, ...)

\method{vec_cast}{bioseq_dna}(x, to, ...)

\method{vec_ptype2}{bioseq_rna}(x, y, ...)

\method{vec_cast}{bioseq_rna}(x, to, ...)
}
\arguments{
\item{x}{an object.}

\item{y}{an object.}

\item{...}{other arguments.}

\item{to}{a class}
}
\description{
Internal

Internal

Internal

Internal

Internal

Internal
}
\keyword{internal}
