% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_aa.R
\name{is_aa}
\alias{is_aa}
\title{Test if the object is an amino acid vector}
\usage{
is_aa(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical.
}
\description{
This function returns TRUE for objects of class bioseq_aa
}
\examples{
x <- c("AGGTGC", "TTCGA")
is_aa(x)
y <- aa(x)
is_aa(x)

}
