% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_dna.R
\name{is_dna}
\alias{is_dna}
\title{Test if the object is a DNA vector}
\usage{
is_dna(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical.
}
\description{
This function returns TRUE for objects of class bioseq_dna
}
\examples{
x <- c("AGGTGC", "TTCGA")
is_dna(x)
y <- dna(x)
is_dna(x)

}
