% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_rna.R
\name{is_rna}
\alias{is_rna}
\title{Test if the object is a RNA vector}
\usage{
is_rna(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical.
}
\description{
This function returns TRUE for objects of class bioseq_rna
}
\examples{
x <- c("AGGTGC", "TTCGA")
is_rna(x)
y <- rna(x)
is_rna(x)

}
