% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_aa.R
\name{pillar_shaft.bioseq_aa}
\alias{pillar_shaft.bioseq_aa}
\title{Internal formatting}
\usage{
pillar_shaft.bioseq_aa(x, ...)
}
\arguments{
\item{x}{a object.}

\item{...}{other params.}
}
\description{
Internal formatting
}
\keyword{internal}
