% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_dna.R
\name{pillar_shaft.bioseq_dna}
\alias{pillar_shaft.bioseq_dna}
\title{Internal formating}
\usage{
pillar_shaft.bioseq_dna(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{other params.}
}
\description{
Internal formating
}
\keyword{internal}
