% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_rna.R
\name{pillar_shaft.bioseq_rna}
\alias{pillar_shaft.bioseq_rna}
\title{Internal formating}
\usage{
pillar_shaft.bioseq_rna(x, ...)
}
\arguments{
\item{x}{a object.}

\item{...}{other params.}
}
\description{
Internal formating
}
\keyword{internal}
