% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_biological_operations.R
\name{rev_complement}
\alias{rev_complement}
\alias{seq_complement}
\alias{seq_reverse}
\title{Reverse and complement sequences}
\usage{
seq_complement(x)

seq_reverse(x)
}
\arguments{
\item{x}{a DNA or RNA vector.
Function \code{seq_reverse} also accepts AA vectors.}
}
\value{
A reverse or complement sequence (same class as the input).
}
\description{
Reverse and complement sequences
}
\examples{
x <- dna("ACTTTGGCTAAG")
seq_reverse(x)
seq_complement(x)

}
\seealso{
Other biological operations: 
\code{\link{seq_rev_translate}()},
\code{\link{seq_translate}()},
\code{\link{transcription}}
}
\concept{biological operations}
