% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_rna.R
\name{rna}
\alias{rna}
\title{Build a RNA vector}
\usage{
rna(...)
}
\arguments{
\item{...}{characters to turn into RNA. Can be a set of name-value pairs.}
}
\value{
a vector of class bioseq_rna
}
\description{
\code{rna()} build a RNA vector from a character vector.
}
\examples{

rna("AGGUGC", "UUCGA")

rna(Seq_1 = "AGGUGC", Seq_2 = "UUCGA")

x <- c("AGGTGC", "TTCGA")
rna(x)

}
\seealso{
Other classes: 
\code{\link{aa}()},
\code{\link{dna}()}
}
\concept{classes}
