% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_stat_operations.R
\name{seq_nseq}
\alias{seq_nseq}
\title{Number of sequences in a vector}
\usage{
seq_nseq(x)
}
\arguments{
\item{x}{a DNA, RNA or AA vector.}
}
\value{
an integer.
}
\description{
This is an alias for \code{length}.
}
\seealso{
Other op-misc: \code{\link{seq_disambiguate_IUPAC}},
  \code{\link{seq_nchar}}, \code{\link{seq_spellout}},
  \code{\link{seq_stat_gc}}, \code{\link{seq_stat_prop}}
}
\concept{op-misc}
