% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_string_operation.R
\name{seq_split_kmer}
\alias{seq_split_kmer}
\title{Split sequences into k-mers}
\usage{
seq_split_kmer(x, k)
}
\arguments{
\item{x}{A DNA, RNA or AA vector.}

\item{k}{an integer giving the size of the k-mer.}
}
\value{
a list of k-mer vectors of same class as \code{x}.
}
\description{
Split sequences into k-mers
}
\examples{

x <- dna("ACGTTAGTGTAGCCGT", "CTCGAAATGA")
seq_split_kmer(x, k = 5)
}
\seealso{
\code{\link{seq_split_pattern}}.
}
