% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_biological_operations.R
\name{transcription}
\alias{transcription}
\alias{seq_transcribe}
\alias{seq_rev_transcribe}
\title{Transcribe DNA, reverse-transcribe RNA}
\usage{
seq_transcribe(x)

seq_rev_transcribe(x)
}
\arguments{
\item{x}{A vector of DNA for \code{seq_transcribe},
a vector of RNA for \code{seq_rev_transcribe}}
}
\value{
A vector of RNA for \code{seq_transcribe},
a vector of DNA for \code{seq_rev_transcribe}
}
\description{
Transcribe DNA, reverse-transcribe RNA
}
