% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences in FASTA format}
\usage{
write_fasta(x, file, append = FALSE, line_length = 80, block_length = 10)
}
\arguments{
\item{x}{a DNA, RNA or AA vector.}

\item{file}{a path to a file or a connection.}

\item{append}{a logical. If \code{TRUE} append the data to the file.
If \code{FALSE} (default), overwrite the file.}

\item{line_length}{length (in number of character) of one line
(excluding spaces separating blocks). Use \code{Inf} to avoid line breaks.}

\item{block_length}{length (in number of character) of one block.
Use the same value as \code{line_length} or \code{Inf} to avoid
block separation.}
}
\description{
Write sequences in FASTA format
}
\seealso{
Other input/output operations: 
\code{\link{read_fasta}()}
}
\concept{input/output operations}
