% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_wrappers.R
\name{aliview}
\alias{aliview}
\title{AliView: DNA sequences viewer}
\usage{
aliview(x, aliview_exec = options("bioseq.aliview.exec"))
}
\arguments{
\item{x}{a DNA, RNA or AA vector.
Alternatively a \code{DNAbin} or \code{AAbin} object.}

\item{aliview_exec}{a character string giving the path of the program.}
}
\description{
This function uses AliView (Larsson, 2014) to visualize DNA sequences.
The software must be installed on the computer.
}
\details{
By default, the function assumes that the executable is installed
in a directory located on the PATH. Alternatively the user can provide
an absolute path to the executable (i.e. the location where the software
was installed/uncompressed). This information can be stored in the global
options settings using
\code{options(bioseq.aliview.exec = "my_path_to_aliview")}.
}
\references{
Larsson, A. (2014). AliView: a fast and lightweight alignment
viewer and editor for large data sets. Bioinformatics30(22): 3276-3278.
}
\seealso{
Other GUI wrappers: \code{\link{seaview}}
}
\concept{GUI wrappers}
