% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences in FASTA format}
\usage{
write_fasta(x, file, append = FALSE)
}
\arguments{
\item{x}{a DNA, RNA or AA vector.}

\item{file}{a path to a file or a connection.}

\item{append}{a logical. If \code{TRUE} append the data to the file.
If \code{FALSE} (default), overwrite the file.}
}
\description{
Write sequences in FASTA format
}
\seealso{
Other io: \code{\link{read_fasta}}
}
\concept{io}
